third_party/WebKit/LayoutTests/inspector/initial-modules-load-expected.txt

This test validates initial set of loaded modules.


Running: testInitialLoad
Loaded modules:
    main
    elements

Running: testCreateElementsPanel
Loaded modules:
    main
    elements

Running: testCreateNetworkPanel
Loaded modules:
    main
    elements
    network
    source_frame

Running: testShowSourcesPanel
Loaded modules:
    main
    elements
    network
    sources
    source_frame

Running: testOpenUISourceCode
Loaded modules:
    main
    elements
    network
    sources
    source_frame
